feat: convert to typescript
- Converts to typescript
- Only named exports
- No more CJS, only ESM
- Runs tests on all supported environments
- Adds auto-publish
- Adds dependabot

BREAKING CHANGE: switch to named exports, ESM only
