--- .travis.yml
@@ -1,9 +1,0 @@
-sudo: required
-language: node_js
-
-addons:
- chrome: stable
-
-node_js:
-  - node
-  - '8'
