+++ package.json
@@ -1,14 +1,10 @@
 {
   "name": "observable-webworkers",
   "version": "1.0.0",
-  "description": "A way to evesdrop on Web Worker messages",
-  "main": "lib/index.js",
-  "scripts": {
-    "test": "ava",
-    "lint": "standard"
-  },
+  "description": "A way to eavesdrop on Web Worker messages",
   "author": "Alex Potsides <alex@achingbrain.net>",
-  "license": "ISC",
+  "license": "Apache-2.0 OR MIT",
+  "homepage": "https://github.com/achingbrain/observable-webworkers#readme",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/achingbrain/observable-webworkers.git"
@@ -16,18 +12,119 @@
   "bugs": {
     "url": "https://github.com/achingbrain/observable-webworkers/issues"
   },
-  "homepage": "https://github.com/achingbrain/observable-webworkers",
-  "devDependencies": {
-    "ava": "^0.25.0",
-    "browserify": "^16.2.2",
-    "execa": "^0.10.0",
-    "pre-commit": "^1.2.2",
-    "run-headless": "^2.0.1",
-    "standard": "^11.0.1",
-    "webworkify": "^1.5.0"
+  "engines": {
+    "node": ">=16.0.0",
+    "npm": ">=7.0.0"
   },
-  "pre-commit": [
-    "lint",
-    "test"
-  ]
+  "type": "module",
+  "types": "./dist/src/index.d.ts",
+  "files": [
+    "src",
+    "dist/src",
+    "!dist/test",
+    "!**/*.tsbuildinfo"
+  ],
+  "exports": {
+    ".": {
+      "import": "./dist/src/index.js"
+    }
+  },
+  "eslintConfig": {
+    "extends": "ipfs",
+    "parserOptions": {
+      "sourceType": "module"
+    }
+  },
+  "release": {
+    "branches": [
+      "master"
+    ],
+    "plugins": [
+      [
+        "@semantic-release/commit-analyzer",
+        {
+          "preset": "conventionalcommits",
+          "releaseRules": [
+            {
+              "breaking": true,
+              "release": "major"
+            },
+            {
+              "revert": true,
+              "release": "patch"
+            },
+            {
+              "type": "feat",
+              "release": "minor"
+            },
+            {
+              "type": "fix",
+              "release": "patch"
+            },
+            {
+              "type": "chore",
+              "release": "patch"
+            },
+            {
+              "type": "docs",
+              "release": "patch"
+            },
+            {
+              "type": "test",
+              "release": "patch"
+            },
+            {
+              "scope": "no-release",
+              "release": false
+            }
+          ]
+        }
+      ],
+      [
+        "@semantic-release/release-notes-generator",
+        {
+          "preset": "conventionalcommits",
+          "presetConfig": {
+            "types": [
+              {
+                "type": "feat",
+                "section": "Features"
+              },
+              {
+                "type": "fix",
+                "section": "Bug Fixes"
+              },
+              {
+                "type": "chore",
+                "section": "Trivial Changes"
+              },
+              {
+                "type": "docs",
+                "section": "Trivial Changes"
+              },
+              {
+                "type": "test",
+                "section": "Tests"
+              }
+            ]
+          }
+        }
+      ],
+      "@semantic-release/changelog",
+      "@semantic-release/npm",
+      "@semantic-release/github",
+      "@semantic-release/git"
+    ]
+  },
+  "scripts": {
+    "build": "tsc",
+    "pretest": "npm run build",
+    "test": "aegir test -t browser -f ./dist/test",
+    "test:chrome": "npm run test",
+    "test:firefox": "npm run test -- -- --browser firefox",
+    "lint": "aegir lint"
+  },
+  "devDependencies": {
+    "aegir": "^36.1.3"
+  }
 }
