+++ tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "aegir/src/config/tsconfig.aegir.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "emitDeclarationOnly": false,
+    "module": "ES2020"
+  },
+  "include": [
+    "src",
+    "test"
+  ]
+}
